// src/components/models/roomDetails.ts
export interface RoomDetailsModel{
                id: number;          
                topic: string;
                cardImage: string;
                guests: string;
                areaOfRoom: string;
                features: string[];
                amentiesIconName: string[];
                price: number;
           
}